Respawn Trial - Marco
The Respawn almost works, I gifured out taht it doesnt capture the location for when game starts. It only captures the rotation for some reason, so it only rotates it back to how it was in the beginning of the game instead of respawning it with its initial rotation and location. I will fix this tomorrow hopefully.