What I did - Marco
Dispenser, collection, and the object absorbing the helium works now. I just need to make the code to make the sped slower the more helium the player has. The amount of helium is displayed on the top left of the string and how much the star has. The star will be destroyed once the amount needed is met.